refactor(cli): consolidate command implementations into core modules
Move command implementations from `impl` submodules into their respective
core modules. This change simplifies the codebase by removing the
`impl` directory and integrating the command logic directly into the
main CLI modules. The `run_command` utility is now part of the core
module, allowing for a more streamlined and maintainable structure.

This refactor reduces complexity by eliminating the need for separate
implementation files, making it easier to manage and understand the
codebase. It also enhances maintainability by centralizing command
execution logic, which can now be reused across different CLI commands.
